+++ bundles/framework/personaldata/Flyout.js
@@ -113,11 +113,30 @@ Oskari.clazz.define('Oskari.mapframework.bundle.personaldata.Flyout',
                 panel,
                 notLoggedIn = this.instance.getLocalization('notLoggedIn'),
                 notLoggedInText = this.instance.getLocalization('notLoggedInText'),
-                notLoggedInFullText = notLoggedIn + ' ' + notLoggedInText;
+                notLoggedInFullText = notLoggedIn,
+                conf = this.instance.conf,
+                lang = Oskari.getLang();
+
+
+            if(conf.logInUrl) {
+                if(typeof conf.logInUrl === 'object') {
+                    var value = conf.logInUrl[lang];
+                    if(!value) {
+                        value = conf.logInUrl[Oskari.getDefaultLanguage()];
+                    }
+                    
+                    if(value) {
+                        notLoggedInText = '<a href="' + value + '">' + notLoggedInText + '</a>';    
+                    }
+                }
+                else if(typeof conf.logInUrl === 'string') { 
+                    notLoggedInText = '<a href="' + conf.logInUrl + '">' + notLoggedInText + '</a>';
+                }
+            }
+
+            notLoggedInFullText += '<br/><br/>' + notLoggedInText;
 
             flyout.empty();
-
-
             
             this.tabsContainer = Oskari.clazz.create('Oskari.userinterface.component.TabContainer', 
                 notLoggedInFullText);